Added configutil package
