+++ cmd/configuration/configuration.go
@@ -6,10 +6,11 @@ import (
 	"io"
 	"os"
 
+	"github.com/flyteorg/flytectl/pkg/configutil"
+
 	initConfig "github.com/flyteorg/flytectl/cmd/config/subcommand/config"
 	cmdcore "github.com/flyteorg/flytectl/cmd/core"
 	cmdUtil "github.com/flyteorg/flytectl/pkg/commandutils"
-	"github.com/flyteorg/flytectl/pkg/util"
 	"github.com/flyteorg/flytestdlib/config/viper"
 	"github.com/manifoldco/promptui"
 
@@ -60,32 +61,32 @@ func configInitFunc(ctx context.Context, args []string, cmdCtx cmdcore.CommandCo
 
 func initFlytectlConfig(reader io.Reader) error {
 
-	templateValues := util.ConfigTemplateValuesSpec{
+	templateValues := configutil.ConfigTemplateValues{
 		Host:     "dns:///localhost:30081",
 		Insecure: initConfig.DefaultConfig.Insecure,
-		Template: util.GetSandboxTemplate(),
+		Template: configutil.GetSandboxTemplate(),
 	}
 
 	if len(initConfig.DefaultConfig.Host) > 0 {
 		templateValues.Host = fmt.Sprintf("dns:///%v", initConfig.DefaultConfig.Host)
-		templateValues.Template = util.GetAWSCloudTemplate()
+		templateValues.Template = configutil.GetAWSCloudTemplate()
 		_, result, err := prompt.Run()
 		if err != nil {
 			return err
 		}
 		if result == "GCS" {
-			templateValues.Template = util.GetGoogleCloudTemplate()
+			templateValues.Template = configutil.GetGoogleCloudTemplate()
 		}
 	}
 	var _err error
-	if _, err := os.Stat(util.ConfigFile); os.IsNotExist(err) {
-		_err = util.SetupConfig(util.ConfigFile, templateValues)
+	if _, err := os.Stat(configutil.ConfigFile); os.IsNotExist(err) {
+		_err = configutil.SetupConfig(configutil.ConfigFile, templateValues)
 	} else {
-		if cmdUtil.AskForConfirmation(fmt.Sprintf("This action will overwrite an existing config file at [%s]. Do you want to continue?", util.ConfigFile), reader) {
-			if err := os.Remove(util.ConfigFile); err != nil {
+		if cmdUtil.AskForConfirmation(fmt.Sprintf("This action will overwrite an existing config file at [%s]. Do you want to continue?", configutil.ConfigFile), reader) {
+			if err := os.Remove(configutil.ConfigFile); err != nil {
 				return err
 			}
-			_err = util.SetupConfig(util.ConfigFile, templateValues)
+			_err = configutil.SetupConfig(configutil.ConfigFile, templateValues)
 		}
 	}
