+++ cmd/sandbox/teardown.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/flyteorg/flytectl/pkg/util"
+	"github.com/flyteorg/flytectl/pkg/configutil"
 
 	"github.com/flyteorg/flytectl/pkg/docker"
 
@@ -45,7 +45,7 @@ func tearDownSandbox(ctx context.Context, cli docker.Docker) error {
 			return err
 		}
 	}
-	if err := util.ConfigCleanup(); err != nil {
+	if err := configutil.ConfigCleanup(); err != nil {
 		fmt.Printf("Config cleanup failed. Which Failed due to %v \n ", err)
 	}
 	fmt.Printf("%v %v Sandbox cluster is removed successfully. \n", emoji.Broom, emoji.Broom)
