+++ pkg/configutil/configutil_test.go
@@ -1,0 +1,60 @@
+package configutil
+
+import (
+	"io/ioutil"
+	"os"
+	"testing"
+
+	"github.com/flyteorg/flytectl/pkg/util"
+
+	f "github.com/flyteorg/flytectl/pkg/filesystemutils"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestConfigCleanup(t *testing.T) {
+	_, err := os.Stat(f.FilePathJoin(f.UserHomeDir(), ".flyte"))
+	if os.IsNotExist(err) {
+		_ = os.MkdirAll(f.FilePathJoin(f.UserHomeDir(), ".flyte"), 0755)
+	}
+	_ = ioutil.WriteFile(FlytectlConfig, []byte("string"), 0600)
+	_ = ioutil.WriteFile(Kubeconfig, []byte("string"), 0600)
+
+	err = ConfigCleanup()
+	assert.Nil(t, err)
+
+	_, err = os.Stat(FlytectlConfig)
+	check := os.IsNotExist(err)
+	assert.Equal(t, check, true)
+
+	_, err = os.Stat(Kubeconfig)
+	check = os.IsNotExist(err)
+	assert.Equal(t, check, true)
+	_ = ConfigCleanup()
+}
+
+func TestSetupFlytectlConfig(t *testing.T) {
+	templateValue := ConfigTemplateValues{
+		Host:     "dns:///localhost:30081",
+		Insecure: true,
+		Template: AdminConfigTemplate,
+	}
+	_, err := os.Stat(f.FilePathJoin(f.UserHomeDir(), ".flyte"))
+	if os.IsNotExist(err) {
+		_ = os.MkdirAll(f.FilePathJoin(f.UserHomeDir(), ".flyte"), 0755)
+	}
+	err = util.SetupFlyteDir()
+	assert.Nil(t, err)
+	err = SetupConfig("version.yaml", templateValue)
+	assert.Nil(t, err)
+	_, err = os.Stat("version.yaml")
+	assert.Nil(t, err)
+	check := os.IsNotExist(err)
+	assert.Equal(t, check, false)
+	_ = ConfigCleanup()
+}
+
+func TestAwsConfig(t *testing.T) {
+	assert.Equal(t, AdminConfigTemplate+StorageS3ConfigTemplate, GetAWSCloudTemplate())
+	assert.Equal(t, AdminConfigTemplate+StorageGCSConfigTemplate, GetGoogleCloudTemplate())
+	assert.Equal(t, AdminConfigTemplate+StorageConfigTemplate, GetSandboxTemplate())
+}
