+++ pkg/util/util_test.go
@@ -1,11 +1,7 @@
 package util
 
 import (
-	"io/ioutil"
-	"os"
 	"testing"
-
-	f "github.com/flyteorg/flytectl/pkg/filesystemutils"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -54,50 +50,6 @@ func TestWriteIntoFile(t *testing.T) {
 	})
 }
 
-func TestConfigCleanup(t *testing.T) {
-	_, err := os.Stat(f.FilePathJoin(f.UserHomeDir(), ".flyte"))
-	if os.IsNotExist(err) {
-		_ = os.MkdirAll(f.FilePathJoin(f.UserHomeDir(), ".flyte"), 0755)
-	}
-	_ = ioutil.WriteFile(FlytectlConfig, []byte("string"), 0600)
-	_ = ioutil.WriteFile(Kubeconfig, []byte("string"), 0600)
-
-	err = ConfigCleanup()
-	assert.Nil(t, err)
-
-	_, err = os.Stat(FlytectlConfig)
-	check := os.IsNotExist(err)
-	assert.Equal(t, check, true)
-
-	_, err = os.Stat(Kubeconfig)
-	check = os.IsNotExist(err)
-	assert.Equal(t, check, true)
-	_ = ConfigCleanup()
-}
-
-func TestSetupFlytectlConfig(t *testing.T) {
-	templateValue := ConfigTemplateValuesSpec{
-		Host:     "dns:///localhost:30081",
-		Insecure: true,
-		Template: AdminConfigTemplate,
-	}
-	_, err := os.Stat(f.FilePathJoin(f.UserHomeDir(), ".flyte"))
-	if os.IsNotExist(err) {
-		_ = os.MkdirAll(f.FilePathJoin(f.UserHomeDir(), ".flyte"), 0755)
-	}
-	err = SetupFlyteDir()
-	assert.Nil(t, err)
-	err = SetupConfig("version.yaml", templateValue)
-	assert.Nil(t, err)
-	_, err = os.Stat("version.yaml")
-	assert.Nil(t, err)
-	check := os.IsNotExist(err)
-	assert.Equal(t, check, false)
-	_ = ConfigCleanup()
-}
-
-func TestAwsConfig(t *testing.T) {
-	assert.Equal(t, AdminConfigTemplate+StorageS3ConfigTemplate, GetAWSCloudTemplate())
-	assert.Equal(t, AdminConfigTemplate+StorageGCSConfigTemplate, GetGoogleCloudTemplate())
-	assert.Equal(t, AdminConfigTemplate+StorageConfigTemplate, GetSandboxTemplate())
+func TestSetupFlyteDir(t *testing.T) {
+	assert.Nil(t, SetupFlyteDir())
 }
